Remove bad/duplicated setup advice
The Installation and Getting Started sections were duplicates, though the Installation section didn't mention setting PATH and mentions some ruby scripts that don't seem to work anymore and don't seem to be necessary. This PATH thing confuses me every time I try to do this after a gap of a few months to forget it again. I never see it because it's mentioned after the first mention of "hem server".