gama: common.styl: Correct use of decals images for 10_x.
Copy decals images used for decals-10_x-2_0,1,3.
Just mentioning a filename in this CSS file is enough to make the
image show up as the thumbnail image for the example image,
but when someone clicks on the thumbail to get the full image,
the web page code has no way to know what was in the CSS file.
It guesses the image URL based on the regular pattern. So if we
don't follow the regular pattern then it doesn't work.

I have code in android-galaxyzoo, that I use when adding a survey,
which actually parses the CSS file and finds these problems.